refactor(elements): extract Hittest+Entries sub-modules
Phase 12a §4 of the CreateElementMediator god-module split.
Moves the locator-race + hit-test cluster and the locator-entry
assembly cluster out of CreateElementMediator.ts into dedicated
sub-modules under Mediator/Elements/Create/.

Why
---
After §3 (Scope+Locators) the parent still owned every primitive
between "I have N candidates, now race them" and "I picked a
winner": context fan-out, .nth() expansion, Playwright visible-
wait, elementFromPoint hit-test, skip-link filter, fallback
selection. That cluster owns no business semantics — it is pure
DOM plumbing the mediator orchestrates. Pulling it out clears 386
lines from the parent (now 1798 LoC, target ≤150) and lets the
next commits work on the higher-level resolvers without scrolling
past plumbing.

What
----
- Create/Hittest.ts (new) — `IRaceDiagnostic`, `raceLocators`,
  `raceLocatorsWithHitTest`. Eight private helpers
  (`isAccessibilitySkipLink`, `isElementHitTestable`,
  `isTrulyVisible`, `awaitVisibleIndices`, `hitTestIndices`,
  `probeNotSkipLinkIndex`, `filterOutSkipLinks`, `resolveWinner`)
  stay encapsulated — they have no out-of-cluster callers.
- Create/Entries.ts (new) — `ILocatorEntry`, `buildLocatorEntries`,
  `buildLocatorEntriesAll`. `getAllContexts`,
  `MAX_NTH_PER_LOCATOR`, `expandLocatorToNth`, `IExpandEntryArgs`,
  `entriesFromExpansion`, `expandCandidateEntries`,
  `mapCandidatesToExpansions` stay private to the cluster.
- Create/index.ts — barrel re-exports the six new public symbols
  alongside Scope+Locators.
- CreateElementMediator.ts — drops the extracted 386-line block
  and pulls the six symbols from the barrel. Drops the now-unused
  `buildCandidateLocatorsBase` import (its only caller moved into
  Entries).
- .github/docs-coverage-allowlist.txt — six allowlist entries
  (Reason per symbol) for the new internal exports. They are
  mediator plumbing only; no public scraper consumes them.

Public surface
--------------
lib/index.d.ts + lib/index.d.cts SHA256 BYTE-IDENTICAL to baseline
(both 14823 B, hash 221E1611…AAB055).
lib/index.cjs grew by 974 B from module-init wrappers for the two
new files — implementation reorganisation, not API change.

Verification
------------
- tsc --noEmit: clean
- eslint Mediator/Elements/Create + parent: clean
- prettier: clean
- lint:architecture: clean
- jest Mediator/Elements (30 suites / 270 tests): all pass
- npm run build: success; d.ts hashes match baseline
