+++ src/Scrapers/Pipeline/Mediator/Elements/CreateElementMediator.ts
@@ -35,11 +35,16 @@ import {
 } from './ActionExecutors.js';
 import {
   buildCandidateLocators,
-  buildCandidateLocatorsBase,
+  buildLocatorEntries,
+  buildLocatorEntriesAll,
   CLICK_RACE_TIMEOUT,
   getActivePhase,
   getActiveStage,
+  type ILocatorEntry,
+  type IRaceDiagnostic,
   NO_FORM_ANCHOR,
+  raceLocators,
+  raceLocatorsWithHitTest,
   setActivePhase as setGlobalPhase,
   setActiveStage as setGlobalStage,
 } from './Create/index.js';
@@ -273,392 +278,6 @@ function buildScopeToForm(cache: IFormCache): IElementMediator['scopeToForm'] {
     const mutable = [...candidates];
     return scopeCandidates(cache.selector, mutable);
   };
-}
-
-/**
- * Collect all contexts to search: main page + child iframes.
- * @param page - The Playwright page.
- * @returns Array of Page/Frame contexts to build locators from.
- */
-function getAllContexts(page: Page): (Page | Frame)[] {
-  const mainFrame = page.mainFrame();
-  const childFrames = page.frames().filter((f): boolean => f !== mainFrame);
-  return [page, ...childFrames];
-}
-
-/** Playwright element wait state for locator races. */
-type WaitState = 'visible' | 'attached';
-
-/**
- * Race all locators in parallel — first matching state wins. Returns winning index or -1.
- * @param locators - Array of Playwright locators to race.
- * @param timeout - Timeout in ms for each locator.
- * @param state - Element state to wait for (default: 'visible').
- * @returns Index of first matching locator, or -1 if none.
- */
-async function raceLocators(
-  locators: Locator[],
-  timeout: number,
-  state: WaitState = 'visible',
-): Promise<number> {
-  const waiters = locators.map(async (loc, i): Promise<number> => {
-    await loc.waitFor({ state, timeout });
-    return i;
-  });
-  const results = await Promise.allSettled(waiters);
-  const winner = results.find((r): boolean => r.status === 'fulfilled');
-  if (winner?.status !== 'fulfilled') return -1;
-  return winner.value;
-}
-
-/**
- * Hit-test: check if the browser compositor can reach this element.
- * Uses elementFromPoint at the element's center — handles ALL hiding:
- * CSS 3D backface, ng-hide inheritance, z-index, overflow clip.
- * @param locator - The Playwright locator to test.
- * @returns True if the element is hit-testable at its center.
- */
-/**
- * Browser-evaluated predicate: returns true iff the element's className
- * contains any well-known accessibility skip-link / sr-only marker
- * (`skip-to-main`, `skip-link`, `sr-only`, `visually-hidden`).
- *
- * <p>Top-level pure function (no captured closures) so Playwright's
- * evaluate serialization can transport it into the page context. The
- * regex literal is duplicated inline between this predicate and
- * {@link isElementHitTestable} because Playwright's `evaluate(fn, arg)`
- * cannot transport a regex argument without an `unknown` cast the
- * project ban-list rejects, and a captured top-level const is not
- * serialized into the page context together with the function body.
- *
- * @param el - Target element under test.
- * @returns true iff the element is an a11y skip-link / sr-only wrapper.
- */
-function isAccessibilitySkipLink(el: Element): boolean {
-  const cls = el.className;
-  return /skip-to-main|skip-link|sr-only|visually-hidden/i.test(cls);
-}
-
-/**
- * Browser-evaluated hit-test predicate. MUST be a top-level pure function
- * (no captured closures) so Playwright's evaluate serialization can
- * transport it into the page context.
- *
- * Rejects disabled placeholders BEFORE hit-test (Wix renders a disabled
- * `<button role="link">` over the real link on some bank templates).
- * Also rejects accessibility skip-link / sr-only / visually-hidden
- * wrappers inline (defense in depth for the #309 Discount skip-link
- * collision — covered upstream by {@link filterOutSkipLinks} but
- * duplicated here so a regression of either guard alone still rejects
- * the wrong element).
- * @param el - Target element under test.
- * @returns True when the element is hit-testable at its center point.
- */
-function isElementHitTestable(el: Element): boolean {
-  if (el.hasAttribute('disabled')) return false;
-  if (el.getAttribute('aria-disabled') === 'true') return false;
-  const cls = el.className;
-  if (/skip-to-main|skip-link|sr-only|visually-hidden/i.test(cls)) return false;
-  const rect = el.getBoundingClientRect();
-  const hit = document.elementFromPoint(rect.left + rect.width / 2, rect.top + rect.height / 2);
-  return el === hit || el.contains(hit);
-}
-
-/**
- * Hit-test — check elementFromPoint at element center.
- * @param locator - The Playwright locator to test.
- * @returns True when the element is hit-testable.
- */
-async function isTrulyVisible(locator: Locator): Promise<boolean> {
-  return locator.evaluate(isElementHitTestable).catch((): boolean => false);
-}
-
-/**
- * Race locators then validate winner with elementFromPoint hit-test.
- * If winner fails hit-test, check remaining settled results.
- * Falls back to first Playwright-visible if no hit-test passes.
- * @param locators - Locators to race.
- * @param timeout - Timeout in ms.
- * @returns Index of first truly visible locator, or -1.
- */
-/** Race diagnostic — trace-level detail about what happened. */
-interface IRaceDiagnostic {
-  readonly winner: number;
-  readonly fulfilledCount: number;
-  readonly hitTestPassedCount: number;
-  readonly fulfilledIndices: readonly number[];
-}
-
-/**
- * Race locators then validate with hit-test. Returns diagnostic.
- * @param locators - Locators to race.
- * @param timeout - Timeout in ms.
- * @returns Diagnostic with winner + fulfilled detail.
- */
-/**
- * Await waitFor(visible) on all locators; return indices that resolved.
- * @param locators - Locators to race.
- * @param timeout - Per-locator waitFor timeout.
- * @returns Indices that passed Playwright visibility check.
- */
-async function awaitVisibleIndices(
-  locators: Locator[],
-  timeout: number,
-): Promise<readonly number[]> {
-  const waiters = locators.map(async (loc, i): Promise<number> => {
-    await loc.waitFor({ state: 'visible', timeout });
-    return i;
-  });
-  const results = await Promise.allSettled(waiters);
-  return results
-    .filter((r): boolean => r.status === 'fulfilled')
-    .map((r): number => (r as PromiseFulfilledResult<number>).value);
-}
-
-/**
- * Run hit-test on every fulfilled index; return those that passed.
- * @param locators - All locators (indexed by fulfilled).
- * @param fulfilled - Indices that already passed visibility check.
- * @returns Indices that ALSO passed elementFromPoint hit-test.
- */
-async function hitTestIndices(
-  locators: Locator[],
-  fulfilled: readonly number[],
-): Promise<readonly number[]> {
-  const promises = fulfilled.map(async (idx): Promise<number> => {
-    const isHit = await isTrulyVisible(locators[idx]);
-    return isHit ? idx : -1;
-  });
-  const tests = await Promise.all(promises);
-  return tests.filter((idx): boolean => idx >= 0);
-}
-
-/**
- * Per-locator skip-link probe. Returns the index when the locator's
- * element is NOT a skip-link; returns -1 otherwise. Errors are treated
- * as "not a skip-link" (best-effort filter, do not block on probe
- * failure).
- * @param locators - All locators (indexed by `idx`).
- * @param idx - Index to probe.
- * @returns `idx` when not a skip-link; `-1` otherwise.
- */
-async function probeNotSkipLinkIndex(locators: Locator[], idx: number): Promise<number> {
-  const isSkip = await locators[idx].evaluate(isAccessibilitySkipLink).catch((): boolean => false);
-  return isSkip ? -1 : idx;
-}
-
-/**
- * Filter fulfilled indices to those whose elements are NOT a11y
- * skip-link / sr-only wrappers. Upstream pre-filter for
- * {@link raceLocatorsWithHitTest} that defends BOTH the
- * `hitPassed[0]` primary path AND the `fulfilled[0]` fallback path in
- * {@link resolveWinner} (#309).
- * @param locators - All locators (indexed by fulfilled).
- * @param fulfilled - Indices that passed visibility check.
- * @returns Indices whose elements are NOT accessibility skip-links.
- */
-async function filterOutSkipLinks(
-  locators: Locator[],
-  fulfilled: readonly number[],
-): Promise<readonly number[]> {
-  const probes = fulfilled.map((idx): Promise<number> => probeNotSkipLinkIndex(locators, idx));
-  const results = await Promise.all(probes);
-  return results.filter((idx): boolean => idx >= 0);
-}
-
-/**
- * Race locators then validate winner with elementFromPoint hit-test.
- * If winner fails hit-test, check remaining settled results.
- * Falls back to first Playwright-visible if no hit-test passes.
- * @param locators - Locators to race.
- * @param timeout - Timeout in ms.
- * @returns Diagnostic with winner + fulfilled detail.
- */
-async function raceLocatorsWithHitTest(
-  locators: Locator[],
-  timeout: number,
-): Promise<IRaceDiagnostic> {
-  const fulfilledRaw = await awaitVisibleIndices(locators, timeout);
-  const fulfilled = await filterOutSkipLinks(locators, fulfilledRaw);
-  const hitPassed = await hitTestIndices(locators, fulfilled);
-  return {
-    winner: resolveWinner(hitPassed, fulfilled),
-    fulfilledCount: fulfilled.length,
-    hitTestPassedCount: hitPassed.length,
-    fulfilledIndices: fulfilled,
-  };
-}
-
-/**
- * Pick winner: hit-test winner first, then first visible fallback.
- * Fallback handles overlays (cookie banners) on bank homepages.
- * @param hitPassed - Indices that passed hit-test.
- * @param fulfilled - Indices that passed waitFor visible.
- * @returns Winner index or -1.
- */
-function resolveWinner(hitPassed: readonly number[], fulfilled: readonly number[]): number {
-  if (hitPassed.length > 0) return hitPassed[0];
-  if (fulfilled.length > 0) return fulfilled[0];
-  return -1;
-}
-
-/** A locator paired with the candidate and context that produced it. */
-interface ILocatorEntry {
-  readonly locator: Locator;
-  readonly candidate: SelectorCandidate;
-  readonly context: Page | Frame;
-}
-
-/**
- * Build locator entries with metadata for all contexts × candidates.
- * Preserves which candidate and context produced each locator.
- * @param page - The Playwright page.
- * @param candidates - WellKnown selector candidates.
- * @param formAnchor - Optional CSS form selector for descendant scoping
- *   applied uniformly to every candidate kind via Locator chaining.
- * @returns Array of locator entries with metadata.
- */
-function buildLocatorEntries(
-  page: Page,
-  candidates: readonly SelectorCandidate[],
-  formAnchor = NO_FORM_ANCHOR,
-): ILocatorEntry[] {
-  const contexts = getAllContexts(page);
-  return contexts.flatMap((ctx): ILocatorEntry[] =>
-    candidates.flatMap((c): ILocatorEntry[] =>
-      buildCandidateLocators(ctx, c, formAnchor).map(
-        (locator): ILocatorEntry => ({ locator, candidate: c, context: ctx }),
-      ),
-    ),
-  );
-}
-
-/** Maximum nth-matches enumerated per locator in `resolveAllVisible` —
- *  surfaces multi-element matches (legacy + modern nav buttons) into the
- *  candidate list without exploding the race work. */
-const MAX_NTH_PER_LOCATOR = 5;
-
-/**
- * Expand a single locator into up to N nth-locators, one per matching DOM
- * element. Uses `locator.count()` to discover how many siblings match,
- * caps at `max`. Used by `resolveAllVisible` to surface legacy/modern
- * sibling pairs (same aria-label, distinct ids) so identity-dedup can
- * collapse race winners that point to the same element while preserving
- * distinct ones.
- * @param base - Base locator (without `.first()`).
- * @param max - Cap on returned nth-locators.
- * @returns Up to `max` `.nth(i)` locators (empty when count() rejects or 0).
- */
-async function expandLocatorToNth(base: Locator, max: number): Promise<readonly Locator[]> {
-  const total = await base.count().catch((): number => 0);
-  if (total <= 0) return [];
-  const limit = Math.min(total, max);
-  const out: Locator[] = [];
-  for (let i = 0; i < limit; i = i + 1) {
-    const nthLocator = base.nth(i);
-    out.push(nthLocator);
-  }
-  return out;
-}
-
-/**
- * Bundled args for the per-candidate locator-entry expander — keeps the
- * helper inside the 3-param ceiling.
- */
-interface IExpandEntryArgs {
-  readonly ctx: Page | Frame;
-  readonly candidate: SelectorCandidate;
-  readonly maxPerLocator: number;
-  readonly formAnchor: string;
-}
-
-/**
- * Expand all base locators for one (context, candidate) pair into nth-
- * enumerated entries. Drops `.first()` semantics by reusing the same
- * builder output (`.nth(i)` chains compose with `.first()` such that
- * `.first()` is `.nth(0)` — Playwright treats them identically).
- * @param args - Bundled context + candidate + per-locator cap + formAnchor.
- * @returns Locator entries (one per nth-match per base locator).
- */
-/**
- * Flatten expanded locators into `ILocatorEntry` records.
- * @param expanded - Per-base expansion of locator nth-matches.
- * @param candidate - The candidate that produced these locators.
- * @param ctx - Page or Frame they belong to.
- * @returns Locator entries (locator + candidate + context).
- */
-function entriesFromExpansion(
-  expanded: readonly (readonly Locator[])[],
-  candidate: SelectorCandidate,
-  ctx: Page | Frame,
-): readonly ILocatorEntry[] {
-  return expanded.flatMap((locs): readonly ILocatorEntry[] =>
-    locs.map((locator): ILocatorEntry => ({ locator, candidate, context: ctx })),
-  );
-}
-
-/**
- * Build locator entries that may surface multiple nth-matches per base
- * locator. Used by resolveAllVisible / multi-match resolvers.
- * @param args - Bundled context + candidate + per-locator cap + formAnchor.
- * @returns Locator entries (one per nth-match per base locator).
- */
-async function expandCandidateEntries(args: IExpandEntryArgs): Promise<readonly ILocatorEntry[]> {
-  const bases = buildCandidateLocatorsBase(args.ctx, args.candidate, args.formAnchor);
-  const expansionPromises = bases.map(
-    (b): Promise<readonly Locator[]> => expandLocatorToNth(b, args.maxPerLocator),
-  );
-  const expanded = await Promise.all(expansionPromises);
-  return entriesFromExpansion(expanded, args.candidate, args.ctx);
-}
-
-/**
- * Build locator entries that surface MULTIPLE elements per locator (up to
- * `MAX_NTH_PER_LOCATOR`). Used only by `resolveAllVisible` so other
- * resolvers (login/preLogin/etc.) keep their `.first()`-only semantics —
- * zero behavioural change for banks that already pass on attempt 0.
- * @param page - Playwright page.
- * @param candidates - WK selector candidates.
- * @param formAnchor - Optional CSS form selector — when set, all candidate
- *   kinds are scoped to descendants of the form via Locator chaining.
- * @returns Locator entries (contexts × candidates × nth-matches).
- */
-async function buildLocatorEntriesAll(
-  page: Page,
-  candidates: readonly SelectorCandidate[],
-  formAnchor = NO_FORM_ANCHOR,
-): Promise<readonly ILocatorEntry[]> {
-  const contexts = getAllContexts(page);
-  const expansionPromises = contexts.flatMap((ctx): Promise<readonly ILocatorEntry[]>[] =>
-    mapCandidatesToExpansions(ctx, candidates, formAnchor),
-  );
-  const groups = await Promise.all(expansionPromises);
-  return groups.flat();
-}
-
-/**
- * Build the per-(ctx,candidate) expansion promises array. Extracted from
- * `buildLocatorEntriesAll` so the parent function's nesting stays inside
- * the depth-1 ceiling.
- * @param ctx - Playwright context (Page or Frame).
- * @param candidates - WK selector candidates to enumerate per locator.
- * @param formAnchor - Optional CSS form selector for descendant scoping.
- * @returns One promise of locator entries per candidate, in input order.
- */
-function mapCandidatesToExpansions(
-  ctx: Page | Frame,
-  candidates: readonly SelectorCandidate[],
-  formAnchor = NO_FORM_ANCHOR,
-): Promise<readonly ILocatorEntry[]>[] {
-  return candidates.map(
-    (c): Promise<readonly ILocatorEntry[]> =>
-      expandCandidateEntries({
-        ctx,
-        candidate: c,
-        maxPerLocator: MAX_NTH_PER_LOCATOR,
-        formAnchor,
-      }),
-  );
 }
 
 /**
